+++ packages/gatsby-dev-cli/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.7.9](https://github.com/gatsbyjs/gatsby/compare/gatsby-dev-cli@2.7.8...gatsby-dev-cli@2.7.9) (2020-06-09)
+
+**Note:** Version bump only for package gatsby-dev-cli
+
 ## [2.7.8](https://github.com/gatsbyjs/gatsby/compare/gatsby-dev-cli@2.7.7...gatsby-dev-cli@2.7.8) (2020-06-02)
 
 ### Bug Fixes