+++ packages/gatsby-plugin-typography/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.5.4](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-typography@2.5.3...gatsby-plugin-typography@2.5.4) (2020-06-09)
+
+**Note:** Version bump only for package gatsby-plugin-typography
+
 ## [2.5.3](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-typography@2.5.2...gatsby-plugin-typography@2.5.3) (2020-06-02)
 
 **Note:** Version bump only for package gatsby-plugin-typography